requirements.txt
# === Production deps ===
Flask>=3.0,<4.0
Flask-SQLAlchemy>=3.1,<4.0
Flask-Login>=0.6.3,<1.0
Flask-Migrate>=4.0,<5.0
SQLAlchemy>=2.0,<3.0
psycopg2-binary>=2.9,<3.0
gunicorn>=21.2,<22.0
python-dotenv>=1.0,<2.0
openai>=1.40,<2.0

# === Tests (shipped to Railway too; ~4 MB, not worth a second file for Phase 0) ===
pytest>=8.0,<9.0
